Remove prints from localisation, still has problem

+++ Project/modules/localisation/src/Localisation.hpp
@@ -10,7 +10,10 @@ extern int Localisation_main();
 extern void probability(int cell_column, int cell_row, double cell_dist, double obj_dist);
 extern void breshams_alg(int i, double sen_hori[], double sen_vert[]);
 extern void Print_Occupancy_Map();
-extern void Digital_Differential_Analyzer(double object_cell_m, double object_cell_n, double tank_cell_m, double tank_cell_n);
+extern void Digital_Differential_Analyzer(double object_cell_m,
+                                          double object_cell_n,
+                                          double tank_cell_m,
+                                          double tank_cell_n);
 
 class Localisation {
 public:
@@ -27,7 +30,7 @@ public:
     static const int start_column;
     static double area;
     static double gridspace;
-    static float map[Grid_m][Grid_m];
+    static double map[Grid_m][Grid_m];
 
     static std::vector<std::pair<int, int>> cell_list;
 };
